IMU data now gets scaled to real-world units, and multiple predict steps are used if the timestamp between two IMU measurements are too far apart.

+++ src/ArgusController/IMUData.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 
 struct IMUData {
-	int16_t aX, aY, aZ;
-	int16_t gX, gY, gZ;
+	float aX, aY, aZ;
+	float gX, gY, gZ;
 	uint64_t timestamp_us; // microseconds
 };
